Into the fire - Prevent enemies resetting health
For Into the Fire the health of a Triton and Dread attacker were being
reset back to full in the time between when they were spawning and
getting sent to attack the base. This was very noticable for the Dread
that spawns at portalsouth since it was very close to the base location.