Updated DLL code: Latest updates to ProductionHelper and added exposed BuildAngleObject() to Lua.
Function Params:

Handle BuildAngleObject(string ODFName, int Team, Position, string Label,  number Angle, number Group, bool SpawnEmpty, bool StartDeploy, number HeightOffset, bool SpreadOut, bool AlignToTerrain, bool AlignUseWater)

ODF: The odf to spawn, just like BuildObject().
Team: The Team #, just like BuildObject().
Pos: The Position, can accept any type, just like BuildObject().
The rest of the parameters are optional, with defaults shown:
Label = "": First it looks for an object with this label to already exist on the map, if it does, it returns that Handle and skips the rest. If it doesn't, it spawns that object with the settings desired.
Angle = -1: The horizontal angle, in degrees, in world coordinates (0 = North) that the object spawns. 90.0 is East, 180.0 is South, 270.0 is West, etc... If left < 0, it doesn't change any angles and spawns things like normal. (BuildObject() using path points with multiple points will spawn objects on point 0, facing the next point, 1)
Group = -1: The Group this object is placed in, if left -1 it uses SetBestGroup().
Empty = false: If this is true, it spawns the object without a pilot in it.
StartDeploy = false: If this is true, it triggers Deploy() as soon as it's spawned.
HeightOffset = 0.0: This is a flat height adjustment done immediately after the object is spawned. Useful for spawning things in the air.
SpreadOut = false (true if Pos is a path point): If this is true, it spreads the units out using GetPositionNear(Pos, 2.0f, 10.0f).
AlignToTerrain = false: If this is true, it does a little more work to GetTerrainHeightAndNormal() and builds a Matrix based on said Normal, to align the object with the terrain.
AlignUseWater = false: This is the "use Water" argument passed into GetTerrainHeightAndNormal() if AlignToTerrain is true.